Move device_info into its own file
